+++ src/commands/registerCommands.ts
@@ -1,4 +1,4 @@
-﻿import * as vscode from "vscode";
+import * as vscode from "vscode";
 import { WorkspaceStore } from "../core/workspaceStore";
 import { ActionsProvider } from "../providers/actionsProvider";
 import { NotesProvider } from "../providers/notesProvider";
@@ -68,7 +68,7 @@ export function registerCommands(
   registerSafeCommand(context, "sshWorkspace.editSavedCommand", (input) => editSavedCommand(store, views, input));
   registerSafeCommand(context, "sshWorkspace.duplicateSavedCommand", (input) => duplicateSavedCommand(store, views, input));
   registerSafeCommand(context, "sshWorkspace.deleteSavedCommand", (input) => deleteSavedCommand(store, views, input));
-  registerSafeCommand(context, "sshWorkspace.runSavedCommand", (input) => runSavedCommand(store, input));
+  registerSafeCommand(context, "sshWorkspace.runSavedCommand", (input) => runSavedCommand(store, views, input));
   registerSafeCommand(context, "sshWorkspace.openSystemStatus", () => openSystemStatus(store));
   registerSafeCommand(context, "sshWorkspace.openFile", (input) => openTrackedFile(store, views, input));
   registerSafeCommand(context, "sshWorkspace.trackCurrentFile", () => trackCurrentFile(store, views));